requirements.txt
flask
flask-jsonpify
flask-sqlalchemy
flask-restful
colorama
pyyaml
pyprocessor

